Algoritmo Generacion beta

+++ Dupicadora/Funciones.h
@@ -9,6 +9,6 @@ void copiaCarga(TFT_eSPI tft);
 void actualizarCarga(TFT_eSPI tft, int progreso);
 void finalizado(TFT_eSPI tft, TFT_eSPI_Button Home);
 void Copiar();
-void generarLlave2(uint16_t key[]);
+void generarLlave2(double key[]);
 void generarLlave(uint16_t key[]);
 void generarLlave(uint16_t key[], uint16_t offsets[5]);
